Add batched stream
