+++ blocks/datasets/text.py
@@ -1,7 +1,9 @@
 import os
 
+import numpy
+
 from blocks import config
-from blocks.datasets import Dataset
+from blocks.datasets import Dataset, CachedDataStream
 
 
 class TextFileState(object):
@@ -178,3 +180,64 @@ class OneBillionWord(TextFile):
                 'news.en.heldout-{:05d}-of-00050'.format(partition))
                 for partition in which_partitions]
         super(OneBillionWord, self).__init__(files, dictionary, **kwargs)
+
+
+class NGramStream(CachedDataStream):
+    """Return n-grams from a stream.
+
+    This data stream wrapper takes as an input a data stream outputting
+    batches of sentences. From these sentences n-grams of a fixed order
+    (e.g. bigrams, trigrams, etc.) are extracted and returned. It also
+    creates a ``targets`` data source. For each example, the target is the
+    word immediately following that n-gram. It is normally used for
+    language modelling, where we try to predict the next word from the
+    previous n words.
+
+    Parameters
+    ----------
+    ngram_order : int
+        The order of the n-grams to output e.g. 3 for trigrams.
+    data_stream : :class:`DataStream` instance
+        The data stream providing sentences. Each example is assumed to be
+        a list of integers.
+    target_source : str, optional
+        This data stream adds a new source for the target words. By default
+        this source is 'targets'.
+
+    Notes
+    -----
+    This class inherits from :class:`CachedDataStream` because it makes use
+    of a cache to store the sentences from the wrapped data stream in.
+
+    """
+    def __init__(self, ngram_order, data_stream, target_source='targets',
+                 iteration_scheme=None):
+        if len(data_stream.sources) > 1:
+            raise ValueError
+        super(NGramStream, self).__init__(data_stream, iteration_scheme)
+        self.sources = self.sources + (target_source,)
+        self.ngram_order = ngram_order
+
+    def get_data(self, request=None):
+        if not self.cache[0]:
+            self._cache()
+        features, targets = [], []
+        for i, sentence in enumerate(self.cache[0]):
+            for j in range(request):
+                features.append(sentence[j:j + self.ngram_order])
+                targets.append([sentence[j + self.ngram_order]])
+                if j + self.ngram_order == len(sentence) - 1:
+                    sentence_ended = True
+                    break
+                elif len(features) == request:
+                    sentence_ended = False
+                    break
+            if sentence_ended:
+                self.cache[0].pop(0)
+                if not self.cache[0]:
+                    self._cache()
+            else:
+                self.cache[0][0] = self.cache[0][0][j + 1:]
+            if len(features) == request:
+                break
+        return tuple(numpy.asarray(data) for data in (features, targets))
